Adding and removing editors to collections

+++ flask_application/controllers/frontend.py
@@ -82,14 +82,15 @@ def unfollow(type, id):
 
 	user = User.objects(id=current_user.get_id()).first()
 	model.remove_follower(user)
-	cached = Cache.objects(name="collections-for-%s" % user).first()
-	if cached:
-		cached.delete()
+	if type=='collection':
+		cached = Cache.objects(name="collections-for-%s" % current_user.get_id()).first()
+		if cached:
+			cached.delete()
 	return jsonify({
 		'result': 'success',
 		'message': get_template_attribute('frontend/macros.html', 'follow')(model)
 	})
-
+	
 
 @frontend.route('/search')
 @frontend.route('/search/<type>')
